# docs/swagger.yaml
openapi: 3.0.0
info:
  title: DevConnect API
  version: 1.0.0
  description: API for managing developers, projects, and tools with JWT authentication
servers:
  - url: /
    description: Current server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    User:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        provider: { type: string, enum: [local, google] }
        avatar: { type: string }
        roles: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
...
paths:
  /auth/register:
    post:
      summary: Register a new local user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
              required: [email, password]
      responses:
        '201':
          description: User created, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string, description: JWT token }
        '409': { description: Email already registered }
        '400': { description: Email and password required }
  /auth/login:
    post:
      summary: Local login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
              required: [email, password]
      responses:
        '200':
          description: Login successful, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '401': { description: Invalid credentials }
...
  /auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Not authenticated }
...